test: check for files before deleting

+++ test/main.js
@@ -1,6 +1,7 @@
 const {readFile, unlink} = require('fs').promises
 const {relative} = require('path').posix
 const glob = require('fast-glob')
+const {existsSync} = require('fs')
 const JSZip = require('jszip')
 const mockArgv = require('mock-argv')
 const t = require('tap')
@@ -9,7 +10,10 @@ const main = require('../lib/main')
 const ARCHIVE_PATH = 'test/fixtures/out/archive.zip'
 
 t.test('--src only', async t => {
-  await unlink(ARCHIVE_PATH)
+  if (existsSync(ARCHIVE_PATH)) {
+    await unlink(ARCHIVE_PATH)
+  }
+
   await mockArgv(
     ['--archive', ARCHIVE_PATH, '--src', 'test/fixtures/in/dir1'],
     async () => {
@@ -36,7 +40,10 @@ t.test('--src only', async t => {
 })
 
 t.test('--root and --src', async t => {
-  await unlink(ARCHIVE_PATH)
+  if (existsSync(ARCHIVE_PATH)) {
+    await unlink(ARCHIVE_PATH)
+  }
+
   await mockArgv(
     [
       '--archive',
